Deploy script
This relies on the makefile for some logical dependency between steps.

The reality is just that I explored 2 different and quick angles with a
Makefile and a script.

One provide dependencies but is less user firendly, the other require a
lot of plumbing internally.

Other tools, such as ansible, saltstack, etc... would have proven
probably like a big hammer for a small nail.